Implement proper `APIClient` mock
The mock makes it easy to verify request and simulate responses.
